Refactor cukes to not include link
Because the rules for generation of the link were hardcoded in tests and made
changes to the Git-related code difficult.

The generation of the links is also unit tested, so adding expectations
in the integration tests did not add much value, especially at the cost
of the technical debt added, and it not being a core business value.

Signed-off-by: Dennis Ideler <dennis.ideler@fundingcircle.com>