Keep it simple for looping through all git repositories to update
The optimization where we reduce the set of repositories to update
was not working as we intended and didn't offer a significant speed
boost.

In its current state, it takes about 40-60 seconds to loop through 37
repositories. We can look at performance again when it becomes an issue.

One idea is to use memoization to speed up the loop by storing the
results of expensive calls and returning the cached result when the same
inputs occur again. For example, avoid re-instantiating all the objects
every time (e.g. GitRepositoryLocation, Rugged::Repository,
GitRepository) by storing 'app_name' => Rugged::Repository pairs.