Swap OctokitClient with GithubClient
The new class was needed because OctokitClient was a singleton, but
we needed two clients because we use different tokens for different
use cases.

The rename signals that we don't care what specific client we use to
interact with GitHub, because we may swap it at some point. We only care
that it's a GitHub client, and that's now clear.