Add GitRepositoryLocation#full_repo_name
To be used when creating a GitHub commit status.

Instead of creating an Octokit::Repository from its GitHub URL (which
we get by converting the URI) and then passing that repository to the
octokit client, it's simpler and less error prone to identify the repo
by using the full repo name as a string.

Once the switch happens, we can finally upgrade Octokit (see #81).
This was not previously possible because we're abusing the
Octokit::Repository.from_url method.