Also strip whitespace from URI when getting params in controller
Unfortunately 1dae31b20ba704370140623e81b0dcfce44d0be9 was not enough,
because the validation steps require the whitespace to be stripped
and the 'before create' hook to strip whitespace runs after our
form validations.

We could also give the user an error if they added whitespace, but
I find that a bit too harsh if the URI is valid besides the whitespace.