Tighter Git URI validation around ports
Now using two regexes instead of one because the cases are slightly
different and easier to handle as two different cases.

If the URI starts with 'git@' then it must end with '.git', and the
content in between is simpler to validate than other URI protocols.

Also note that the repo name section of the URI has been tightened up
as well. For example, it can no longer be 'owner/repo~name' because
such a name would be converted to 'owner/repo-name' by services such
as GitHub.