Repo is not accessible if format is incorrect
A URI like 'https://github.com/foo' will pass regex validation (because
it is a valid Git URI) but blows up when given to Octokit because GitHub
follow the owner/repo format.

Now we treat such URIs as the repository being not accessible, and
present an error message instead of the app blowing up.

Signed-off-by: Loga Jegede <loga.jegede@fundingcircle.com>