Only snapshot tickets if they contain a Feature Review
If JIRA is configured to send webhook events for any Issue activity,
then a massive amount of ticket snapshots will unnecessarily be created.

There is no need to prematurely snapshot tickets without a Feature
Review because many of them will never contain a Feature Review.

Once we do get a ticket event for a Feature Review being linked, then
that same payload contains all the other information we need to create
the snapshots, such as ticket key, title (summary), description, etc.

Note that this breaks the other specs. To be fixed in the next commit!