:fire: Remove unused VersionResolver
Was used only by the FR search, for finding relevant commits to include
in the search. No longer necessary because we only look at the most
recent FR now, which is either at the head of the topic branch or
at the merge commit or second parent of the merge commit.