Update alert messaging to be more accurate
When we try to expand a version in the DeployEvent model, but the lookup
fails, then the commit will be nil, which triggers a DeployAlert.

In these cases it's more accurate to say that the version is unknown
instead of missing, because the payload might have had some gibberish
version.