Remove test that was not adding value
We were testing that a Date object is set to the beginning and end of
a day, but it would pass even if we swapped those expectations around
because a Date object does not contain detailed information about time
of day. Note that in the implementation we also do not set the Date to
the beginning or end of day.