Differentiate between ticket unapproval and rejection
Unapproval = Ticket was approved (passed review) and later unapproved
Reject = Ticket was rejected before being approved (in review stage)

We were using 'rejected' for an unapproval event in our JiraEvent
factory which was confusing. It has been renamed to 'unapproved'.

Also move statuses in the JiraEvent factory to their own traits, so we
can set the status of a ticket without having to go through a status
change, which happens in cases such as a comment event.