Set approved_at for approvals or carry over for approved tickets
Previous commit introduced a bug where it did not carryover the approval
time for tickets that already have a snapshot and are approved (e.g.
moving from 'ready for deploy' to 'done' should keep approval time).

Note that with these changes the approved_at time will always be
correctly set as the approval time, since we restrict the initial
setting to be for approval events.