Add repo owners
Repo owners are added from an edit page for git_repository_locations
The repo owners are also shown on the index page for repositories

In the future they will be able to specify a comment will prevent
deployment alert if there isn't a ticket and they'll receive a email for
deployment alerts.

You can link to existing repo owners with only email:
  test@foo.com - will search for such project owner or create one
  John Snow <test@foo.com> - will search with the email and create / update
    the project owner by setting also the name

Flow:

```
GitRepositoryLocationsController#edit / #update
  |
  --> EditGitRepositoryLocationForm#call
        |
        --> create Events::RepoOwnershipEvent
        --> apply event Repositories::RepoOwnershipRepository
              |
              --> create / update Snapshots::RepoOwnership
              --> create / update RepoOwner
```

Events::RepoOwnershipEvent
  - app_name
  - repo_owners - list of emails (string)
  - email - of the current user
  - timestamps

RepoOwner - keeps the name of owners in sync between repositories
  - name
  - email
  - timestamps

MailAddressList - helper for handling list of emails

Events for Repositories::RepoOwnershipRepository are not updated in the background.
They are applied immediately when the form is submitted. This is done in
hackish way. Maybe if the repositories update the current EventCount
things will be better.

When running the recreate job the events are applied also to Repositories::RepoOwnershipRepository