Make sure that we have only events that can be applied by repositories
When applying ReleaseExceptionEvent we validate that the repo owner has
access to the repo. When we recreate the database this will fail.

The validation should happen before we create the event so we can
recreate the database easily without having to deal with invalid events.

Modify the EventsController so it can handle simple validations
Factories::EventFactory now builds a new event without saving it, so you
can save it on your own and handle any errors