Export the repositories that do not run in the background into config/initializers/repositories.rb
These repositories are applied by the updater only when recreating the
database. This means that they are manually applied in the context of
the Repositories::Updater

At the moment there is only one such repository Repositories::RepoOwnershipRepository which
applies the events when a form is submitted which prevents the weird
messages that tell users that it will take a little bit of time before
their action takes place.