Prevent repo re-clone caused by concurrent fetches
The git repositories are updated by both the git-worker process and the
RelinkTicketJob delayed-job.

libgit2 creates a lock file when it updates a reference in a git fetch
operation, e.g.
  /tmp/6-funding_circle_app/.git/refs/remotes/origin/di-ri-928.lock.

If this fetch operation is in progress when a libgit2 fetch from the
other process tries to update the same reference, it causes an exception

Rugged::OSError: failed to create locked file
 '/tmp/6-funding_circle_app/.git/refs/remotes/origin/di-ri-928.lock':
 File exists

We currently re-clone the repository when there is any Rugged::OSError,
but this particular one is recoverable without doing a clone, which can
take a long time for a large repository.

This will instead retry a git fetch up to 10 times if a concurrent git
fetch has a lock on the same reference.

It sleeps for 1 second before each retry.

See https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/5497