When restoring the gem cache fallback to an incomplete cache if no exact match
When we restore the bundled rubygems cache we currently only check for a match on the checksum of the Gemfile.lock. This is not optimal because you can still gain major performance improvements from caching gems even if your cache only contains part of the required dependencies.

As such when restoring the cache if the job fails to find an exact match for the gemfile.lock checksum it will now fall back to the latest cached gemset.

Read more about circleci cache restoration at https://circleci.com/docs/2.0/caching/#restoring-cache